Ready for Review

+++ src/utils/api.js
@@ -12,6 +12,12 @@ export const getItems = () =>
   fetch(`${baseUrl}/items`, { headers }).then(handleServerResponse);
 
 export const addItem = ({ name, imageUrl, weather }) => {
+  const token = localStorage.getItem("jwt");
+
+  if (!token) {
+    return;
+  }
+
   return fetch(`${baseUrl}/items`, {
     method: "POST",
     headers: {
@@ -27,8 +33,46 @@ export const addItem = ({ name, imageUrl, weather }) => {
 };
 
 export const removeItem = (itemId) => {
+  const token = localStorage.getItem("jwt");
+
+  if (!token) {
+    return;
+  }
+
   return fetch(`${baseUrl}/items/${itemId}`, {
     method: "DELETE",
     headers,
+  }).then(handleServerResponse);
+};
+
+export const likeItem = (itemId) => {
+  const token = localStorage.getItem("jwt");
+
+  if (!token) {
+    return;
+  }
+
+  return fetch(`${baseUrl}/items/${itemId}/likes`, {
+    method: "PUT",
+    headers: {
+      "Content-type": "application/json",
+      authorization: `Bearer ${token}`,
+    },
+  }).then(handleServerResponse);
+};
+
+export const unlikeItem = (itemId) => {
+  const token = localStorage.getItem("jwt");
+
+  if (!token) {
+    return;
+  }
+
+  return fetch(`${baseUrl}/items/${itemId}/likes`, {
+    method: "DELETE",
+    headers: {
+      "Content-type": "application/json",
+      authorization: `Bearer ${token}`,
+    },
   }).then(handleServerResponse);
 };
